Made Subcollection a subclass of Collection and removed the need for the collection_action_mixin.rb
